Move mempool handle to a separate module
