+++ mempool/src/pool.rs
@@ -22,7 +22,6 @@ use std::time::Duration;
 use common::chain::Block;
 use common::primitives::BlockHeight;
 use tokio::sync::mpsc;
-use tokio::sync::oneshot;
 
 use chainstate::chainstate_interface::ChainstateInterface;
 use common::chain::signed_transaction::SignedTransaction;
@@ -50,6 +49,7 @@ use crate::error::TxValidationError;
 use crate::feerate::FeeRate;
 use crate::feerate::INCREMENTAL_RELAY_FEE_RATE;
 use crate::feerate::INCREMENTAL_RELAY_THRESHOLD;
+use crate::handle::MempoolMethodCall;
 use crate::tx_accumulator::TransactionAccumulator;
 use crate::MempoolEvent;
 use store::MempoolRemovalReason;
@@ -795,91 +795,6 @@ impl GetMemoryUsage for SystemUsageEstimator {
     fn get_memory_usage(&self) -> MemoryUsage {
         //TODO implement real usage estimation here
         0
-    }
-}
-
-pub struct MempoolInterfaceHandle {
-    sender: mpsc::UnboundedSender<MempoolMethodCall>,
-}
-
-impl MempoolInterfaceHandle {
-    pub fn new(sender: mpsc::UnboundedSender<MempoolMethodCall>) -> Self {
-        Self { sender }
-    }
-}
-
-pub enum MempoolMethodCall {
-    AddTransaction {
-        tx: SignedTransaction,
-        rtx: oneshot::Sender<Result<(), Error>>,
-    },
-    GetAll {
-        rtx: oneshot::Sender<Vec<SignedTransaction>>,
-    },
-    CollectTxs {
-        tx_accumulator: Box<dyn TransactionAccumulator + Send>,
-        rtx: oneshot::Sender<Box<dyn TransactionAccumulator>>,
-    },
-    ContainsTransaction {
-        tx_id: Id<Transaction>,
-        rtx: oneshot::Sender<bool>,
-    },
-    SubscribeToEvents {
-        handler: Arc<dyn Fn(MempoolEvent) + Sync + Send>,
-        rtx: oneshot::Sender<()>,
-    },
-}
-
-#[async_trait::async_trait]
-impl MempoolInterface for MempoolInterfaceHandle {
-    async fn add_transaction(&mut self, tx: SignedTransaction) -> Result<(), Error> {
-        let (rtx, rrx) = tokio::sync::oneshot::channel();
-        self.sender
-            .send(MempoolMethodCall::AddTransaction { tx, rtx })
-            .map_err(|_| Error::SendError)?;
-        rrx.await.map_err(|_| Error::RecvError)?
-    }
-
-    async fn get_all(&self) -> Result<Vec<SignedTransaction>, Error> {
-        let (rtx, rrx) = tokio::sync::oneshot::channel();
-        self.sender
-            .send(MempoolMethodCall::GetAll { rtx })
-            .map_err(|_| Error::SendError)?;
-        rrx.await.map_err(|_| Error::RecvError)
-    }
-
-    // Returns `true` if the mempool contains a transaction with the given id, `false` otherwise.
-    async fn contains_transaction(&self, tx_id: &Id<Transaction>) -> Result<bool, Error> {
-        let (rtx, rrx) = tokio::sync::oneshot::channel();
-        self.sender
-            .send(MempoolMethodCall::ContainsTransaction { tx_id: *tx_id, rtx })
-            .map_err(|_| Error::SendError)?;
-        rrx.await.map_err(|_| Error::RecvError)
-    }
-
-    async fn collect_txs(
-        &self,
-        tx_accumulator: Box<dyn TransactionAccumulator + Send>,
-    ) -> Result<Box<dyn TransactionAccumulator>, Error> {
-        let (rtx, rrx) = tokio::sync::oneshot::channel();
-        self.sender
-            .send(MempoolMethodCall::CollectTxs {
-                tx_accumulator,
-                rtx,
-            })
-            .map_err(|_| Error::SendError)?;
-        rrx.await.map_err(|_| Error::RecvError)
-    }
-
-    async fn subscribe_to_events(
-        &mut self,
-        handler: Arc<dyn Fn(MempoolEvent) + Send + Sync>,
-    ) -> Result<(), Error> {
-        let (rtx, rrx) = tokio::sync::oneshot::channel();
-        self.sender
-            .send(MempoolMethodCall::SubscribeToEvents { handler, rtx })
-            .map_err(|_| Error::SendError)?;
-        rrx.await.map_err(|_| Error::RecvError)
     }
 }
